only create files if already they dont exist

+++ docker-entrypoint.sh
@@ -1,33 +1,36 @@
 #!/bin/sh
+if [ ! -f stunnel.pem ]; then
 
-touch openssl.cnf
+    cat >> openssl.cnf <<_EOF_
+    [ req ]
+    prompt = no
+    distinguished_name = req_distinguished_name
 
-cat >> openssl.cnf <<_EOF_
-[ req ]
-prompt = no
-distinguished_name = req_distinguished_name
-
-[ req_distinguished_name ]
-CN = ${CN}
+    [ req_distinguished_name ]
+    CN = ${CN}
 _EOF_
 
-openssl req -x509 -nodes -new -newkey rsa:2048 -days 3650 -config openssl.cnf -keyout stunnel.pem -out stunnel.pem
-chmod 600 stunnel.pem
-rm openssl.cnf
+    openssl req -x509 -nodes -new -newkey rsa:2048 -days 3650 -config openssl.cnf -keyout stunnel.pem -out stunnel.pem
+    chmod 600 stunnel.pem
+    rm openssl.cnf
+fi
 
-cat > stunnel.conf <<_EOF_
-foreground = yes
-setuid = stunnel4
-setgid = stunnel4
-socket = l:TCP_NODELAY=1
-socket = r:TCP_NODELAY=1
-cert = /stunnel/stunnel.pem
-client = ${CLIENT:-no}
-pid = /var/run/stunnel4/stunnel.pid
+if [ ! -f stunnel.conf ]; then
 
-[${SERVICE}]
-accept = ${ACCEPT}
-connect = ${CONNECT}
+    cat > stunnel.conf <<_EOF_
+    foreground = yes
+    setuid = stunnel4
+    setgid = stunnel4
+    socket = l:TCP_NODELAY=1
+    socket = r:TCP_NODELAY=1
+    cert = /stunnel/stunnel.pem
+    client = ${CLIENT:-no}
+    pid = /var/run/stunnel4/stunnel.pid
+
+    [${SERVICE}]
+    accept = ${ACCEPT}
+    connect = ${CONNECT}
 _EOF_
+fi
 
 exec stunnel "stunnel.conf"
